Create separate component for article view.
